requirements.txt
numpy==1.23.5
pandas==2.1.1
openai-whisper # última versión (versión 20231117 es la instalada por psc)